requirements.txt
mistralai==0.4.2
chainlit==2.5.5
pypdf
python-docx
python-dotenv